Added mtl to the /lib/ directory. Hopefully, this will be a useful matrix library.
git-svn-id: https://gorion.oerc.ox.ac.uk:8443/svn/spinach_gui@46 fbaf8dc8-a59a-9f45-831a-ff29167b0329

+++ wxSpinach/gui/glMolDisplay.cpp
@@ -79,8 +79,8 @@ void glMolDisplay::glTick() {
 //    glLoadMatrixf(rotationMatrix);
 
 
-    long spinCount=mParentFrame->mSpinSys->GetSpinCount();
-    for(long i=0; i<spinCount ;i++) {
+    unsigned spinCount=mParentFrame->mSpinSys->GetSpinCount();
+    for(unsigned long i=0; i<spinCount ;i++) {
         if(i==mParentFrame->GetFocusedSpin()) {
             glColor3f(1.0, 1.0, 1.0);
         } else {
